@@ -6,7 +6,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.19.3 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   flutter:
@@ -15,7 +15,7 @@ dependencies:
   get_it: ^7.2.0
   flutter_bloc: ^8.1.2
   font_awesome_flutter: ^10.4.0
-  intl: ^0.17.0
+  intl: ^0.18.0
   cached_network_image: ^3.2.3
   fluttertoast: ^8.2.1
   pin_code_fields: ^8.0.1
@@ -34,38 +34,38 @@ dependencies:
   esc_pos_printer: ^4.1.0
   image: ^3.0.8
   esc_pos_utils: ^1.1.0
-  flutter_local_notifications: ^15.1.0+1
-  permission_handler: ^10.4.3
+  flutter_local_notifications: ^16.2.0
+  permission_handler: ^11.0.1
   shimmer: ^3.0.0
   shared_preferences: ^2.0.8
   bloc: ^8.0.1
   easy_localization: ^3.0.1
   expandable_bottom_sheet: ^1.1.1+1
   flutterme_credit_card: ^1.0.2
-  easy_stepper: ^0.6.0
-  go_router: ^10.0.0
-  bottom_sheet: ^3.1.2
+  easy_stepper: ^0.8.3
+  go_router: ^12.1.3
+  bottom_sheet: ^4.0.0
   logger: ^2.0.1
   animate_do: ^3.0.2
   wave: ^0.2.2
   flutter_onboarding_slider: ^1.0.9
-  google_fonts: ^4.0.0
+  google_fonts: ^6.1.0
   calendar_builder: ^0.0.6
   flutter_screenutil: ^5.6.0
   flutter_spinkit: ^5.2.0
-  flutter_datetime_picker: ^1.5.1
   url_launcher: ^6.1.4
   custom_date_range_picker: ^1.0.6
   calendar_date_picker2: ^0.5.3
   multiple_images_picker: ^1.0.1
   image_picker: ^1.0.4
   google_maps_flutter: ^2.2.5
-  geolocator: ^9.0.2
+  geolocator: ^10.1.0
   share_plus: ^7.2.1
+  dropdown_button2: ^2.3.9
 
 
   #connectivity
-  connectivity_plus: ^4.0.2
+  connectivity_plus: ^5.0.2
 
   #firebase
 #  firebase_messaging: ^14.6.5
@@ -81,7 +81,7 @@ dependencies:
   animated_switcher_plus: ^1.0.3
   carousel_slider: ^4.2.1
   gallery_image_viewer: ^1.2.0
-  galleryimage: ^1.2.0
+  galleryimage: ^2.0.1
   flutter_staggered_grid_view: ^0.7.0
   swipe_image_gallery: ^0.8.3
   paged_vertical_calendar: ^1.1.6
@@ -91,8 +91,14 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.0
+  flutter_lints: ^3.0.1
 
+
+dependency_overrides:
+  modal_bottom_sheet:
+    git:
+      url: https://github.com/danReynolds/modal_bottom_sheet.git
+      path: modal_bottom_sheet
 # The following section is specific to Flutter packages.
 flutter:
   uses-material-design: true
