@@ -1,6 +1,3 @@
-import 'dart:math';
-
-// import 'package:custom_date_range_picker/custom_date_range_picker.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:calendar_date_picker2/calendar_date_picker2.dart';
 import 'package:easy_localization/easy_localization.dart';
@@ -12,23 +9,20 @@ import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:go_router/go_router.dart';
 import 'package:kashtat/Core/Cubit/AppCubit.dart';
 import 'package:kashtat/Core/Cubit/AppState.dart';
-import 'package:kashtat/Core/Edited%20Packages/customDateRangePicker.dart';
 import 'package:kashtat/Core/constants/ColorManager.dart';
 import 'package:kashtat/Core/constants/FontManager.dart';
-import 'package:kashtat/Core/constants/ImageManager.dart';
 import 'package:kashtat/Features/Home%20Screen/Widgets/EmptyWidget.dart';
 import 'package:kashtat/Features/Home%20Screen/Widgets/FilterButtonWidget.dart';
 import 'package:kashtat/Features/Home%20Screen/Widgets/TypeWidget.dart';
 import 'package:kashtat/Features/Widgets/HeaderWidget.dart';
 import 'package:kashtat/Features/Widgets/Loader.dart';
 import 'package:kashtat/Features/Widgets/kButton.dart';
-import 'package:kashtat/translations/locale_keys.g.dart';
 
 import '../Request Details/Widgets/ShowModelBottomSheet.dart';
 import 'Widgets/HomeItemWidget.dart';
 
 class HomeScreen extends StatefulWidget {
-  const HomeScreen({Key? key, required this.title}) : super(key: key);
+  const HomeScreen({super.key, required this.title});
   final String title;
 
   @override
@@ -51,7 +45,7 @@ class _HomeScreenState extends State<HomeScreen> {
       // print(controller.position);
       // print(controller.position.pixels);
 
-      if(controller.position.pixels > 50){
+      if(controller.position.pixels > 40){
         setState(() {
           isHorizontal = true;
         });
@@ -90,7 +84,7 @@ class _HomeScreenState extends State<HomeScreen> {
                 final date2 = DateTime.now();
                 final difference = birthday.difference(date2).inDays;
                 if(difference>0){
-                  print('cancel');
+                  // print('cancel');
                   return;
                 }
                 customBottomSheet(
@@ -116,15 +110,15 @@ class _HomeScreenState extends State<HomeScreen> {
                                         child: CachedNetworkImage(
                                           imageUrl: cubit.lastTrip?.unit.mainPic??'',
                                           fit: BoxFit.cover,
-                                          placeholder: (context, url) => Loader(),
-                                          errorWidget: (context, url, error) => Icon(Icons.image_not_supported_outlined,color: Colors.grey,),
+                                          placeholder: (context, url) => const Loader(),
+                                          errorWidget: (context, url, error) => const Icon(Icons.image_not_supported_outlined,color: Colors.grey,),
 
                                         ),
                                       ),
                                     ),
-                                    SizedBox(height: 20,),
-                                    Text('كيف كانت الكشتة مع',style: TextStyle(color: Colors.black,fontSize: 16),),
-                                    SizedBox(height: 10,),
+                                    const SizedBox(height: 20,),
+                                    const Text('كيف كانت الكشتة مع',style: TextStyle(color: Colors.black,fontSize: 16),),
+                                    const SizedBox(height: 10,),
                                     Text(cubit.lastTrip?.unit.name??'',style: TextStyle(color: ColorManager.mainlyBlueColor,fontWeight: FontWeight.bold,fontSize: 16),),
                                     Row(
                                       mainAxisAlignment: MainAxisAlignment.center,
@@ -137,14 +131,14 @@ class _HomeScreenState extends State<HomeScreen> {
                                                 innerSetState((){
                                                   rating = i+1;
                                                 });
-                                                print(rating);
+                                                // print(rating);
                                               },
                                               child: FaIcon(FontAwesomeIcons.solidStar,color:i<rating? Colors.amber :  Colors.grey,),
                                             ),
                                           )
                                       ],
                                     ),
-                                    SizedBox(height: 20,),
+                                    const SizedBox(height: 20,),
 
                                     Row(
                                       mainAxisAlignment: MainAxisAlignment.spaceEvenly,
@@ -196,16 +190,10 @@ class _HomeScreenState extends State<HomeScreen> {
     super.initState();
   }
 
-
-  @override
-  void didChangeDependencies() {
-    super.didChangeDependencies();
-  }
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body:!mounted?SizedBox(): BlocBuilder<AppBloc, AppState>(
+      body:BlocBuilder<AppBloc, AppState>(
         builder: (context, state) {
           final cubit = BlocProvider.of<AppBloc>(context);
           return BlocConsumer<AppBloc, AppState>(
@@ -262,187 +250,191 @@ class _HomeScreenState extends State<HomeScreen> {
                           ),
                         ),
                         const SizedBox(height: 10),
-                        if(isHorizontal)
-                          SizedBox(
-                            height: 40,
-                            child: ListView.builder(
-                              physics: const BouncingScrollPhysics(),
-                              scrollDirection: Axis.horizontal,
-                              itemCount: cubit.allCategories.length,
-                              padding: const EdgeInsets.all( 0.0),
-                              itemBuilder: (context,index){
-                                return InkWell(
-                                  onTap: (){
-                                    cubit.filterHomeUnitsByCategory(catId: cubit.allCategories[index].id??0);
-                                    setState(() {
-                                      selected = cubit.allCategories[index].id??0;
-                                      selectedSub = -1;
-                                    });
-                                  },
-                                  child: Padding(
-                                    padding: const EdgeInsets.symmetric(horizontal: 5.0),
-                                    child: Container(
-                                      decoration: BoxDecoration(
-                                        color: selected == cubit.allCategories[index].id
-                                            ? ColorManager.orangeColor
-                                            : ColorManager.mainlyBlueColor,
-                                        borderRadius: BorderRadius.circular(5)
-                                      ),
+                        AnimatedCrossFade(
+                          duration: const Duration(milliseconds: 350),
+                          firstChild: Padding(
+                            padding: const EdgeInsets.only(bottom: 10.0),
+                            child: SizedBox(
+                              height: 40,
+                              child: ListView.builder(
+                                  physics: const BouncingScrollPhysics(),
+                                  scrollDirection: Axis.horizontal,
+                                  itemCount: cubit.allCategories.length,
+                                  padding: const EdgeInsets.all( 0.0),
+                                  itemBuilder: (context,index){
+                                    return InkWell(
+                                      onTap: (){
+                                        cubit.filterHomeUnitsByCategory(catId: cubit.allCategories[index].id??0);
+                                        setState(() {
+                                          selected = cubit.allCategories[index].id??0;
+                                          selectedSub = -1;
+                                        });
+                                      },
                                       child: Padding(
-                                        padding: const EdgeInsets.symmetric(horizontal: 15.0),
-                                        child: Center(
-                                          child: Text(cubit.allCategories[index].name??'',style: TextStyle(
-                                            fontSize: FontSize.s14,
-                                            fontWeight: FontWeightManager.bold,
-                                            color: Colors.white,
-                                          ), textAlign: TextAlign.center,),
+                                        padding: const EdgeInsets.symmetric(horizontal: 5.0),
+                                        child: Container(
+                                          decoration: BoxDecoration(
+                                              color: selected == cubit.allCategories[index].id
+                                                  ? ColorManager.orangeColor
+                                                  : ColorManager.mainlyBlueColor,
+                                              borderRadius: BorderRadius.circular(5)
+                                          ),
+                                          child: Padding(
+                                            padding: const EdgeInsets.symmetric(horizontal: 15.0),
+                                            child: Center(
+                                              child: Text(cubit.allCategories[index].name??'',style: TextStyle(
+                                                fontSize: FontSize.s14,
+                                                fontWeight: FontWeightManager.bold,
+                                                color: Colors.white,
+                                              ), textAlign: TextAlign.center,),
+                                            ),
+                                          ),
                                         ),
                                       ),
-                                    ),
-                                  ),
-                                );
-                              }),
+                                    );
+                                  }),
+                            ),
                           ),
+                          secondChild: Column(
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              ListView.builder(
+                                  physics: const NeverScrollableScrollPhysics(),
+                                  shrinkWrap: true,
+                                  itemCount: cubit.categoriesWithSubCategories.length,
+                                  padding: const EdgeInsets.all( 0.0),
+                                  itemBuilder: (context,index){
+                                    return Padding(
+                                      padding: const EdgeInsets.only(bottom: 10.0),
+                                      child: Column(
+                                        mainAxisAlignment: MainAxisAlignment.start,
+                                        crossAxisAlignment: CrossAxisAlignment.start,
+                                        children: [
+                                          HomeTypeWidget(
+                                            title: cubit.categoriesWithSubCategories[index].name??'',
+                                            img: cubit.categoriesWithSubCategories[index].img??'',
+                                            color: selected == cubit.categoriesWithSubCategories[index].id
+                                                ? ColorManager.orangeColor
+                                                : null,
+                                            onTap: () {
+                                              cubit.filterHomeUnitsByCategory(catId: cubit.categoriesWithSubCategories[index].id??0);
+                                              setState(() {
+                                                selected = cubit.categoriesWithSubCategories[index].id??0;
+                                                selectedSub = -1;
+                                              });
+                                            },
+                                            count: cubit.categoriesWithSubCategories[index].unitsCount.toString(),
+                                            width: MediaQuery.of(context).size.width,
+                                          ),
+                                          if (selected == cubit.categoriesWithSubCategories[index].id)
+                                            const SizedBox(height: 10),
+                                          if (selected == cubit.categoriesWithSubCategories[index].id)
+                                            GridView.custom(
+                                              primary: false,
+                                              padding: const EdgeInsets.all(0),
+                                              physics: const NeverScrollableScrollPhysics(),
+                                              shrinkWrap: true,
+                                              childrenDelegate: SliverChildBuilderDelegate(
+                                                childCount: cubit.categoriesWithSubCategories[index].subCategories!.length,
+                                                    (context, innerIndex) =>
 
-                        if(isHorizontal)
-                          const SizedBox(height: 10),
-                        if(!isHorizontal)
-                          ListView.builder(
-                              physics: NeverScrollableScrollPhysics(),
-                              shrinkWrap: true,
-                              itemCount: cubit.categoriesWithSubCategories.length,
-                              padding: const EdgeInsets.all( 0.0),
-                              itemBuilder: (context,index){
-                                return Padding(
-                                  padding: const EdgeInsets.only(bottom: 10.0),
-                                  child: Column(
-                                    mainAxisAlignment: MainAxisAlignment.start,
-                                    crossAxisAlignment: CrossAxisAlignment.start,
-                                    children: [
-                                      HomeTypeWidget(
-                                        title: cubit.categoriesWithSubCategories[index].name??'',
-                                        img: cubit.categoriesWithSubCategories[index].img??'',
-                                        color: selected == cubit.categoriesWithSubCategories[index].id
+                                                    InkWell(
+                                                      onTap: () {
+                                                        cubit.filterHomeUnitsByCategory(
+                                                          catId: cubit.categoriesWithSubCategories[index].id??0,
+                                                          subCatId: cubit.categoriesWithSubCategories[index].subCategories![innerIndex].id??0,
+                                                        );
+                                                        setState(() {
+                                                          selectedSub = cubit.categoriesWithSubCategories[index].subCategories![innerIndex].id??0;
+                                                        });
+                                                      },
+                                                      child: Container(
+                                                        decoration: BoxDecoration(
+                                                          borderRadius:
+                                                          BorderRadius.circular(8),
+                                                          border: Border.all(
+                                                              color: selectedSub != cubit.categoriesWithSubCategories[index].subCategories![innerIndex].id
+                                                                  ? ColorManager.orangeColor
+                                                                  : ColorManager.mainlyBlueColor),
+                                                          color: selectedSub == cubit.categoriesWithSubCategories[index].subCategories![innerIndex].id
+                                                              ? ColorManager.mainlyBlueColor
+                                                              : Colors.white,
+                                                        ),
+                                                        child: Center(
+                                                          child: Padding(
+                                                            padding: const EdgeInsets.all(2.0),
+                                                            child: Text(
+                                                              cubit.categoriesWithSubCategories[index].subCategories![innerIndex].name??'',
+                                                              style: TextStyle(
+                                                                color: selectedSub != cubit.categoriesWithSubCategories[index].subCategories![innerIndex].id
+                                                                    ? ColorManager.mainlyBlueColor
+                                                                    : Colors.white,
+
+                                                                fontWeight: FontWeightManager.bold,
+                                                              ),
+                                                              textAlign: TextAlign.center,
+                                                            ),
+                                                          ),
+                                                        ),
+
+                                                      ),
+                                                    ),
+                                              ),
+                                              semanticChildCount: 1,
+                                              gridDelegate: SliverWovenGridDelegate.count(
+                                                crossAxisCount: 4,
+
+                                                mainAxisSpacing: 8,
+                                                crossAxisSpacing: 8,
+                                                pattern: [
+                                                  const WovenGridTile(1.8),
+                                                ],
+                                              ),
+                                            ),
+                                        ],
+                                      ),
+                                    );
+                                  }),
+                              GridView.custom(
+                                primary: false,
+                                padding: const EdgeInsets.all(0),
+                                physics: const NeverScrollableScrollPhysics(),
+                                shrinkWrap: true,
+                                childrenDelegate: SliverChildBuilderDelegate(
+                                  childCount: cubit.categoriesWithoutSubCategories.length,
+                                      (context, index) =>
+                                      HomeTypeWidgetGrid(
+                                        title: cubit.categoriesWithoutSubCategories[index].name??'',
+                                        img: cubit.categoriesWithoutSubCategories[index].img??'',
+                                        bgIconColor: ColorManager.orangeColor,
+                                        color: selected == cubit.categoriesWithoutSubCategories[index].id
                                             ? ColorManager.orangeColor
                                             : null,
                                         onTap: () {
-                                          cubit.filterHomeUnitsByCategory(catId: cubit.categoriesWithSubCategories[index].id??0);
+                                          cubit.filterHomeUnitsByCategory(catId: cubit.categoriesWithoutSubCategories[index].id??0);
                                           setState(() {
-                                            selected = cubit.categoriesWithSubCategories[index].id??0;
-                                            selectedSub = -1;
+                                            selected = cubit.categoriesWithoutSubCategories[index].id??0;
                                           });
                                         },
-                                        count: cubit.categoriesWithSubCategories[index].unitsCount.toString(),
-                                        width: MediaQuery.of(context).size.width,
+                                        count: cubit.categoriesWithoutSubCategories[index].unitsCount.toString(),
                                       ),
-                                      if (selected == cubit.categoriesWithSubCategories[index].id)
-                                        const SizedBox(height: 10),
-                                      if (selected == cubit.categoriesWithSubCategories[index].id)
-                                        GridView.custom(
-                                          primary: false,
-                                          padding: const EdgeInsets.all(0),
-                                          physics: const NeverScrollableScrollPhysics(),
-                                          shrinkWrap: true,
-                                          childrenDelegate: SliverChildBuilderDelegate(
-                                            childCount: cubit.categoriesWithSubCategories[index].subCategories!.length,
-                                                (context, innerIndex) =>
+                                ),
+                                semanticChildCount: 1,
+                                gridDelegate: SliverWovenGridDelegate.count(
+                                  crossAxisCount: 3,
 
-                                                InkWell(
-                                                  onTap: () {
-                                                    cubit.filterHomeUnitsByCategory(
-                                                      catId: cubit.categoriesWithSubCategories[index].id??0,
-                                                      subCatId: cubit.categoriesWithSubCategories[index].subCategories![innerIndex].id??0,
-                                                    );
-                                                    setState(() {
-                                                      selectedSub = cubit.categoriesWithSubCategories[index].subCategories![innerIndex].id??0;
-                                                    });
-                                                  },
-                                                  child: Container(
-                                                    decoration: BoxDecoration(
-                                                      borderRadius:
-                                                      BorderRadius.circular(8),
-                                                      border: Border.all(
-                                                          color: selectedSub != cubit.categoriesWithSubCategories[index].subCategories![innerIndex].id
-                                                              ? ColorManager.orangeColor
-                                                              : ColorManager.mainlyBlueColor),
-                                                      color: selectedSub == cubit.categoriesWithSubCategories[index].subCategories![innerIndex].id
-                                                          ? ColorManager.mainlyBlueColor
-                                                          : Colors.white,
-                                                    ),
-                                                    child: Padding(
-                                                      padding: const EdgeInsets.all(8.0),
-                                                      child: Text(
-                                                        cubit.categoriesWithSubCategories[index].subCategories![innerIndex].name??'',
-                                                        style: TextStyle(
-                                                          color: selectedSub != cubit.categoriesWithSubCategories[index].subCategories![innerIndex].id
-                                                              ? ColorManager.mainlyBlueColor
-                                                              : Colors.white,
-
-                                                          fontWeight: FontWeightManager.bold,
-                                                        ),
-                                                        textAlign: TextAlign.center,
-                                                      ),
-                                                    ),
-
-                                                  ),
-                                                ),
-                                          ),
-                                          semanticChildCount: 1,
-                                          gridDelegate: SliverWovenGridDelegate.count(
-                                            crossAxisCount: 4,
-
-                                            mainAxisSpacing: 8,
-                                            crossAxisSpacing: 8,
-                                            pattern: [
-                                              WovenGridTile(1.6),
-                                            ],
-                                          ),
-                                        ),
-                                    ],
-                                  ),
-                                );
-                              }),
-                        if(!isHorizontal)
-                          Column(
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                            GridView.custom(
-                              primary: false,
-                              padding: const EdgeInsets.all(0),
-                              physics: const NeverScrollableScrollPhysics(),
-                              shrinkWrap: true,
-                              childrenDelegate: SliverChildBuilderDelegate(
-                                childCount: cubit.categoriesWithoutSubCategories.length,
-                                    (context, index) =>
-                                    HomeTypeWidgetGrid(
-                                      title: cubit.categoriesWithoutSubCategories[index].name??'',
-                                      img: cubit.categoriesWithoutSubCategories[index].img??'',
-                                      bgIconColor: ColorManager.orangeColor,
-                                      color: selected == cubit.categoriesWithoutSubCategories[index].id
-                                          ? ColorManager.orangeColor
-                                          : null,
-                                      onTap: () {
-                                        cubit.filterHomeUnitsByCategory(catId: cubit.categoriesWithoutSubCategories[index].id??0);
-                                        setState(() {
-                                          selected = cubit.categoriesWithoutSubCategories[index].id??0;
-                                        });
-                                      },
-                                      count: cubit.categoriesWithoutSubCategories[index].unitsCount.toString(),
-                                    ),
+                                  mainAxisSpacing: 8,
+                                  crossAxisSpacing: 8,
+                                  pattern: [
+                                    const WovenGridTile(1.7),
+                                  ],
+                                ),
                               ),
-                              semanticChildCount: 1,
-                              gridDelegate: SliverWovenGridDelegate.count(
-                                crossAxisCount: 3,
-
-                                mainAxisSpacing: 8,
-                                crossAxisSpacing: 8,
-                                pattern: [
-                                  WovenGridTile(1.7),
-                                ],
-                              ),
-                            ),
-                            const SizedBox(height: 10),
-                          ],
+                              const SizedBox(height: 10),
+                            ],
+                          ),
+                          crossFadeState: isHorizontal ? CrossFadeState.showFirst : CrossFadeState.showSecond,
+                          // crossFadeState: isHorizontal ? CrossFadeState.showFirst : CrossFadeState.showSecond,
                         ),
                       ],
                     ),
@@ -457,7 +449,7 @@ class _HomeScreenState extends State<HomeScreen> {
                         state is LoadingHomeTripsState
                             ? const Loader()
                             : (cubit.homeTrips == null || cubit.homeTrips!.data!.isEmpty)
-                                ? HomeEmptyWidget()
+                                ? const HomeEmptyWidget()
                                 : ListView.builder(
                                     itemBuilder: (context, index) => cubit.homeTrips != null
                                         ? HomeItemWidget(
@@ -506,7 +498,7 @@ class _HomeScreenState extends State<HomeScreen> {
                 selectedDayTextStyle: const TextStyle(color: Colors.white, fontWeight: FontWeight.w700),
                 selectedDayHighlightColor: Colors.purple[800],
                 centerAlignModePicker: true,
-                customModePickerIcon: SizedBox(),
+                customModePickerIcon: const SizedBox(),
                 // dayBuilder: _yourDayBuilder,
                 dayBuilder: ({
                   required date,
@@ -526,7 +518,7 @@ class _HomeScreenState extends State<HomeScreen> {
                           children: [
                             Text(
                               MaterialLocalizations.of(context).formatDecimal(date.day),
-                              style: TextStyle(
+                              style: const TextStyle(
                                 color: Colors.grey,
                                 fontSize: 12
                               ),
@@ -558,7 +550,7 @@ class _HomeScreenState extends State<HomeScreen> {
                 setState(() {
                   _dates = dates;
                 });
-                print(_dates);
+                // print(_dates);
                 if (_dates.isNotEmpty) {
                   if(calculateDifference(_dates.first??DateTime.now())<0){
                     return;
@@ -574,8 +566,8 @@ class _HomeScreenState extends State<HomeScreen> {
                     });
                   }
 
-                  print('----------------------dates-------------------');
-                  print(selectedDates);
+                  // print('----------------------dates-------------------');
+                  // print(selectedDates);
 
                   Navigator.pop(context);
                   final cubit = BlocProvider.of<AppBloc>(context);
@@ -644,7 +636,7 @@ class _HomeScreenState extends State<HomeScreen> {
               child: ListView.builder(
                   shrinkWrap: true,
                   physics:
-                  BouncingScrollPhysics(),
+                  const BouncingScrollPhysics(),
                   itemCount: cubit
                       .allCities.length,
                   itemBuilder:
